+++ src/js/control-bar/progress-control/seek-bar.js
@@ -37,8 +37,8 @@ class SeekBar extends Slider {
     return this.player_.currentTime() / this.player_.duration();
   }
 
-  onMouseDown(event) {
-    super.onMouseDown(event);
+  handleMouseDown(event) {
+    super.handleMouseDown(event);
 
     this.player_.scrubbing(true);
 
@@ -46,7 +46,7 @@ class SeekBar extends Slider {
     this.player_.pause();
   }
 
-  onMouseMove(event) {
+  handleMouseMove(event) {
     let newTime = this.calculateDistance(event) * this.player_.duration();
 
     // Don't let video end while scrubbing.
@@ -56,8 +56,8 @@ class SeekBar extends Slider {
     this.player_.currentTime(newTime);
   }
 
-  onMouseUp(event) {
-    super.onMouseUp(event);
+  handleMouseUp(event) {
+    super.handleMouseUp(event);
 
     this.player_.scrubbing(false);
     if (this.videoWasPlaying) {